preprocessor cleaning
