+++ tests/modelling/test_preprocessor.py
@@ -3,42 +3,14 @@
 import pytest
 from unittest.mock import patch
 import itertools
-import pandas as pd
-from pandas.testing import assert_series_equal, assert_frame_equal
+from pandas.testing import assert_frame_equal
 
 
 from py3dinterpolations.core.griddata import GridData
-from py3dinterpolations.modelling.preprocessor import Preprocessor
 from py3dinterpolations.modelling.preprocessor import (
+    Preprocessor,
     reverse_preprocessing,
-    _standardize,
 )
-
-
-### test _standardize function
-def test_standardize():
-    """test that the standardize function returns the right values"""
-    data = pd.DataFrame({"V": [1, 2, 3, 4, 5, 6, 7, 8, 9]})
-    output = _standardize(data["V"])
-    assert_series_equal(
-        output[0],
-        pd.Series(
-            [
-                -1.460593,
-                -1.095445,
-                -0.730297,
-                -0.365148,
-                0.000000,
-                0.365148,
-                0.730297,
-                1.095445,
-                1.460593,
-            ],
-            name="V",
-        ),
-    )
-    assert isinstance(output[1], dict)
-    assert output[1] == {"mean": 5.0, "std": 2.7386127875258306}
 
 
 def test_Preprocessor_init(test_data):
@@ -51,10 +23,12 @@ def test_Preprocessor_init(test_data):
     )
     # griddata is set correctly
     assert preprocess.griddata == gd
-    # Preprocessor parameters are set correctly
+
+    # Preprocessor Default parameters are set correctly
     assert preprocess.downsampling_res is None
     assert preprocess.standardize_v is True
     assert preprocess.normalize_xyz is True
+
     # preprocessing is not run
     assert preprocess.preprocessor_params == {}
 
@@ -65,49 +39,6 @@ def test_Preprocessor_init(test_data):
 PREPROCESSING_COMBINATIONS = list(
     itertools.product([4, None], [True, False], [True, False])
 )
-
-
-@pytest.mark.parametrize(
-    "downsampling_res, normalize_xyz, standardize_v",
-    PREPROCESSING_COMBINATIONS,
-)
-def test_Preprocessor_preprocess_subcalls(
-    downsampling_res, normalize_xyz, standardize_v, test_data
-):
-    """test that the Preprocessor class calls the right methods"""
-    # this tests uses PREPROCESSING_COMBINATIONS to test all possible
-    # combinations of downsampling_res, normalize_xyz, standardize_v
-    # Create a mock for each of the three methods
-    with patch(
-        "py3dinterpolations.modelling.preprocessor.Preprocessor._downsample_data"
-    ) as mock_downsample, patch(
-        "py3dinterpolations.modelling.preprocessor.Preprocessor._normalize_xyz"
-    ) as mock_normalize, patch(
-        "py3dinterpolations.modelling.preprocessor.Preprocessor._standardize_v"
-    ) as mock_standardize:
-        # Create a Preprocessor object with different arguments
-        gd = GridData(test_data)
-        pp = Preprocessor(
-            gd,
-            downsampling_res=downsampling_res,
-            normalize_xyz=normalize_xyz,
-            standardize_v=standardize_v,
-        )
-
-        # Call the preprocess() method
-        pp.preprocess()
-
-        if downsampling_res is not None:
-            # assert mock_downsample was called once
-            mock_downsample.assert_called_once()
-
-        if normalize_xyz:
-            # assert mock_normalize was called once
-            mock_normalize.assert_called_once()
-
-        if standardize_v:
-            # assert mock_standardize was called once
-            mock_standardize.assert_called_once()
 
 
 @pytest.mark.parametrize(
@@ -129,12 +60,34 @@ def test_Preprocessor_preprocess_output(
         standardize_v=standardize_v,
     ).preprocess()
 
-    # is a griddata object
+    # output is a griddata object
     assert isinstance(pp_gd, GridData)
 
+    # if downsampling_res is not None,
+    # less data is expected
     if downsampling_res is not None:
         # has less data
         assert len(gd.data) > len(pp_gd.data)
+
+    # normalize means that are between 0 and 1
+    if normalize_xyz:
+        # X Y Z are normalized
+        df = pp_gd.data.copy().reset_index()
+        assert df["X"].min() == 0
+        assert df["X"].max() == 1
+        assert df["Y"].min() == 0
+        assert df["Y"].max() == 1
+        assert df["Z"].min() == 0
+        assert df["Z"].max() == 1
+
+    def compare_floats(a, b, eps=1e-10):
+        return abs(a - b) < eps
+    
+    # standardize means that mean is 0 and std is 1
+    if standardize_v:
+        # V is standardized
+        assert compare_floats(pp_gd.data["V"].mean(), 0.0)
+        assert compare_floats(pp_gd.data["V"].std(), 1.0)
 
     # in any case it would not change number of unique ids
     assert len(gd.data.index.get_level_values("ID").unique()) == len(
